Tests re-encryption via re-ingest
If an encrypted AIP is reingested and then stored in an encrypted space
whose GPG key is different from the one originally used to encrypt the
AIP, then the reingested AIP will be encrypted with the new key, i.e.,
the current GPG key of the space. Effectively, AIP re-ingest is the
means to re-encrypt with a different key.