PID Binding Test
Implemented test for PID Binding. Archivematica can be configured to
make requests to a Handle System HTTP API so that files, directories and
entire AIPs can be assigned persistent identifiers (PIDs) and derived
persistent URLs (PURLs). Test committed here creates an AIP and ensures
that its METS file contains the PIDs and PURLs in the correct locations
and that the PURLs are resolvable.