Add feature: performance gains without stdout
This commit introduces feature file performance-stdout-no-write.feature, which
first creates an AIP using a version of Archivematica where the MCP client
passes its client script output streams (stdout and stderr) to the MCP server
for ultimate storage in the database. The timing duration of each client
script run in the course of building that AIP is queried from the
database and saved to disk. The feature then modifies the Archivematica
instance so that the MCP client no longer passes output streams to the
server, and it creates a new AIP, and saves information on micro-service
execution time to disk. Finally, it asserts that the micro-services took
less time in the second case than in the first.

This feature is essentially the executable documentation of a performance
experiment against Archivematica. The current finding is that removing
output streams results in a 6% reduction in task execution time.

Warning: it is a known bug that executing this feature will stall at the
second instance of the `And standard AIP-creation decisions are made`
step.