Modify performance "no stdout" feature
Modifies the performance-stdout-no-write feature file so that it
accounts for the new behaviour of the CAPTURE_CLIENT_SCRIPT_OUTPUT
setting of Archivematica (cf. AM PR 937). Now, when the aforementioned
setting is set to `false`, stdout will never be passed from gearman
worker to gearman task manager, but stderr will be passed only if the
exit code of the task is non-zero. Running this test confirms that the
roughly 6-7% reduction in task execution time is still observed with
these changes.