Add feature for MediaConch-challenging transfer
Adds a feature that demonstrates that Archivematica with MediaConch v.
17.12 installed can correctly validate (and document the validation of) a .mkv
derivative that breaks MediaConch v. 16.12.

Note: this feature implicitly requires a directory in archivematica-sampledata/
called mc-challenging-file/ that contains a single .mp4 file called
lunchroom_manners_512kb.mp4 which, when normalized to .mkv, will produce a file
that makes MediaConch 16.12 but not 17.12 hang when performing an
implementation check. The archive containing this file can be found at
https://github.com/artefactual/archivematica/issues/966.  If possible, the
AM-sampledata repo should be modified to include mc-challenging-file/.