Make timeouts and waits centralized and configurable
Brings all of the constants relevant to timeout and "max wait" logic
into one place (features/environment.py) and makes them customizable via
behave user data.

Fixes #92

Also fixes the downloading of backlogged transfers. Prior to this, the `When
the user waits for the DIP to appear in transfer backlog` step could fail
because the `docker cp` command would be targeting a file that did not yet
exist on the container. This commit increases the 1-second sleep statement to a
3-second one to allow for this. It also adds logging to indicate whether the
timeout was exceeded or whether the transfer was located.

Fixes #86