Remove standard ports from url
Browsers remove the port from the url if it's one of the standard ports. This
commit does the same so the internal representation and the one reported by
the browser match which is important in some places in the code of this project
where the urls are evaluated and compared.

This closes #98.