Fix PID binding acceptance tests
This commit adds a handful of HTML ids required for decision points
to be resolved when testing PID binding. The PID binding test has
been updated to use these constants.

We also fix the retrieval of the AIP preview URL to be consistent
with other URL getters in the code-base.

UUIDs marked as incorrect have been verified as incorrect and
removed.

Auto-approval introduced in Archivematica 1.8 is reflected in the
code. Where the `-D am_version=1.8` flag is used with behave,
approval is not consisdered a mandatory stage when initialising a
transfer.

* Resolves archivematica/issues#243