Correct PID binding tests
In this commit we make a handful of small changes to make the PID
binding tests pass and slowly start moving to making further use of
AMClient API calls. In this instance, we enable the download of
AIP METS from a Archival Information Package without having to read
it via the selenium driver in the Archivematica ingest browser.