Release AMAUAT tests for Archivematica 1.10
As part of the 1.10 release of Archivematica we are releasing a
baseline set of the AMAUAT tests whose properties are well known to
improve the consistency of testing in future.

Archivematica 1.9 and 1.10 also introduced a number of breaking
changes that needed to be rectified with this work.

Namely:

 * PREMIS3 was introduced into Archivematica 1.10.

 * Changes to the processing configuration were introduced as part
   of the format identification configuration changes in
   Archivematica 1.9.

As a result of the black-box tests created with Wellcome Trust, the
1.10 tests now adopt a more API like approach to serve its goal.
This should bring minor speed improvements where decision points are
not being resolved by Selenium interaction with a web-browser. This
should also bring with it greater reliability.

Selected changes:

Where the AMAUAT code attempts multiple retries, e.g. when trying to
access UI components, we adopt a more exponential-back-off like
method to ensure we generate recursion errors before finding what we
need.

Where the code attempts to access METS from the AIP to read its
contents we're almost doing this exclusively through the API now
instead of via an awaiting decision point in the Archivematica
Dashboard.

Similarly, where transfers can be started through the API, we have
changed that mechanism rather than asking Selenium to click around
the interface for us.

Where tests previously relied heavily on resolving decision points
during processing we have made some progress to resolving these at
least via the UI during the test-setup.

Working tests:

Test tags that are now expected to run are:

    * Ingest MKV conformance check feature: @icc
    * Infinite AIP ingest feature: @infinite-aips
    * Metadata-only reingest feature: @mo-aip-reingest
    * Pre-ingest MKV conformance check feature: @picc
    * Transfer MKV policy check feature: @tpc
    * Assign UUIDs to directories feature: @uuids-dirs

Work-in progress tests:

These tests will pass fairly consistently in Docker-compose with a
reasonably performing computer. We have had difficulties running
these on more moderately provisioned virtual machines.

    * AIP encryption feature: @aip-encrypt
    * AIP encryption with replication feature: @aip-encrypt-mirror
    * Ingest MKV policy check feature: @ipc

Co-authored-by: Douglas Cerna (Soy Douglas) <douglascerna@yahoo.com>
Co-authored-by: Santiago Rodríguez <santi@collazo.eu>