Make uuids-dirs tag to extract to a temp directory
The uuids-dirs tag tries to extract the zipped bag test case to the
sample data directory. If the file system is readonly (like in the
compose dev environment) this fails and the rest of the steps lose
track of the files and directories relevant to the transfer.

This fixes it by extracting to a temporary directory instead.